Static withMetaForm and asBodilessLink
