+++ packages/bodiless-components/src/Link/asBodilessLink.tsx
@@ -14,7 +14,6 @@
 
 import React, { ComponentType } from 'react';
 import {
-  useMenuOptionUI,
   asBodilessComponent,
   ifEditable,
   withExtendHandler,
@@ -26,7 +25,6 @@ import {
 } from '@bodiless/core';
 import type { BodilessOptions } from '@bodiless/core';
 import flowRight from 'lodash/flowRight';
-import identity from 'lodash/identity';
 import {
   Fragment,
   addProps,
@@ -36,7 +34,6 @@ import {
   flowIf,
   addClassesIf,
 } from '@bodiless/fclasses';
-import { withFieldApi } from 'informed';
 import { useGetDisabledPages } from '../PageDisable';
 import DefaultNormalHref from './NormalHref';
 import withGoToLinkButton from './withGoToLinkButton';
@@ -49,83 +46,19 @@ import {
   ExtraLinkOptions,
   AsBodilessLink,
 } from './types';
-import { FileUpload as BaseFileUpload, FileUploadProps } from '../FileUpload';
-
-const DEFAULT_INSTRUCTIONS = `
-  Use a fully formed URL only for external links, e.g., https://www.example.com.
-  Internal links should be specified without a protocol or domain. Internal
-  links beginning with a './' will be relative to the current page. Those not
-  beginning with a './' will be prefixed with '/' and be relative to
-  the site root.  All links will have a trailing slash appended.
-`;
-
-const DEFAULT_ALLOWED_FILE_TYPES = [
-  // pdf
-  'application/pdf',
-  // doc
-  'application/msword',
-  // docx
-  'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-];
-
-const FileUpload: ComponentType<Omit<FileUploadProps, 'fieldApi'>> = withFieldApi('href')(BaseFileUpload);
+import useEditFormOverrides from './useEditFormOverrides.bl-edit';
 
 const useLinkOverrides = (useOverrides: UseLinkOverrides = () => ({})): UseLinkOverrides => (
   props => {
     const overrides = useOverrides(props);
     const {
-      submitValueHandler: submitValueHandler$ = identity,
       normalizeHref = (href?: string) => new DefaultNormalHref(href).toString(),
-      instructions = DEFAULT_INSTRUCTIONS,
-      fileUpload = {},
     } = overrides;
-    const { accept: fileUploadAccept = DEFAULT_ALLOWED_FILE_TYPES } = fileUpload;
-    const submitValueHandler = ({ href, 'aria-label': ariaLabel }: LinkData) => submitValueHandler$({
-      href: normalizeHref(href),
-      'aria-label': ariaLabel,
-    });
+    const editFormOverrides = useEditFormOverrides(overrides);
     const finalOverrides: Partial<EditButtonOptions<Props, LinkData>> & ExtraLinkOptions = {
       ...overrides,
       normalizeHref,
-      submitValueHandler,
-      renderForm: ({
-        componentProps: { unwrap, ui: { fileUpload: fileUploadUI } = {} },
-        closeForm,
-      }) => {
-        const {
-          ComponentFormTitle,
-          ComponentFormLabel,
-          ComponentFormText,
-          ComponentFormUnwrapButton,
-          ComponentFormDescription,
-        } = useMenuOptionUI();
-        const removeLinkHandler = (event: React.MouseEvent) => {
-          event.preventDefault();
-          if (unwrap) {
-            unwrap();
-          }
-          closeForm(event);
-        };
-        return (
-          <>
-            <ComponentFormTitle>Link</ComponentFormTitle>
-            <ComponentFormLabel htmlFor="link-href">URL</ComponentFormLabel>
-            <ComponentFormText field="href" id="link-href" aria-describedby="description" placeholder="/link" />
-            <ComponentFormDescription id="description">
-              {instructions}
-            </ComponentFormDescription>
-            <ComponentFormLabel htmlFor="aria-label">Aria Label</ComponentFormLabel>
-            <ComponentFormText field="aria-label" id="aria-label" aria-describedby="description" placeholder="aria-label" />
-            <ComponentFormLabel>File Upload</ComponentFormLabel>
-            <FileUpload ui={fileUploadUI} accept={fileUploadAccept} />
-            {unwrap && (
-            <ComponentFormUnwrapButton type="button" onClick={removeLinkHandler}>
-              Remove Link
-            </ComponentFormUnwrapButton>
-            )}
-          </>
-        );
-      },
+      ...editFormOverrides,
     };
     return finalOverrides;
   }
